Deployment/requirements.txt
flask
python-dotenv
yt-dlp
youtube_transcript_api
langchain
pandas
tqdm
sentence-transformers
pinecone-client
torch
transformers
peft